requirements.txt
discord.py
psutil
iso8601
pytz
pyparsing
cachetools
BeautifulSoup4
pillow
concurrent-log-handler
requests
